+++ app/src/main/res/layout/fragment_manage_tags.xml
