+++ app/src/main/res/layout/fragment_view_item.xml
@@ -252,7 +252,7 @@
                     android:padding="0dp"
                     android:gravity="center_vertical"
                     app:icon="@drawable/baseline_add_24"
-                    app:iconSize="30dp"
+                    app:iconSize="25dp"
                     app:iconTint="@color/black" />
             </LinearLayout>
